+++ sdk/communication/communication-chat/test/utils/recordedClient.ts
@@ -8,7 +8,7 @@ import { env, Recorder, record, RecorderEnvironmentSetup } from "@azure/test-uti
 import { isNode } from "@azure/core-http";
 import { ChatClient } from "../../src";
 import {
-  CommunicationUser,
+  CommunicationUserIdentifier,
   AzureCommunicationUserCredential,
   parseClientArguments
 } from "@azure/communication-common";
@@ -47,7 +47,7 @@ export async function createTestUser(): Promise<CommunicationUserToken> {
   return await identityClient.issueToken(testUser, ["chat"]);
 }
 
-export async function deleteTestUser(testUser: CommunicationUser) {
+export async function deleteTestUser(testUser: CommunicationUserIdentifier) {
   if (testUser) {
     const identityClient = new CommunicationIdentityClient(env.COMMUNICATION_CONNECTION_STRING);
     await identityClient.deleteUser(testUser);