+++ sdk/communication/communication-common/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## 1.0.0-beta.4 (Unreleased)
 
+### Breaking Changes
+
+- Renamed `Identifier` to `CommunicationIdentifier`.
+- Renamed `IdentifierKind` to `CommunicationIdentifierKind`.
+
+- Renamed `CommunicationUser` to `CommunicationUserIdentifier`.
+- Renamed `CommunicationUserKind` to `CommunicationUserIdentifierKind`.
+- Renamed `isCommunicationUser` to `isCommunicationUserIdentifier`.
+
+- Renamed `PhoneNumber` to `PhoneNumberIdentifier`.
+- Renamed `PhoneNumberKind` to `PhoneNumberIdentifierKind`.
+- Renamed `isPhoneNumber` to `isPhoneNumberIdentifier`.
+
+- Renamed `CallingApplication` to `CallingApplicationIdentifier`.
+- Renamed `CallingApplicationKind` to `CallingApplicationIdentifierKind`.
+- Renamed `isCallingApplication` to `isCallingApplicationIdentifier`.
+
 
 ## 1.0.0-beta.3 (2020-11-16)